requirements.txt
asgiref==3.5.0
Django==4.0.2
image==1.5.33
imageio==2.14.1
networkx==2.6.3
numpy==1.22.1
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.0
pyparsing==3.0.7
PyWavelets==1.2.0
pywin32==303
scikit-image==0.19.1
scipy==1.7.3
six==1.16.0
sqlparse==0.4.2
tifffile==2021.11.2
tzdata==2021.5
requests~=2.27.1